add model 14
There is no model 13; this model adds the logbook index with lambda=0. Some derived quants have stderrs of zero, which means SS3 is including values that are constant in the stderr report. Not an issue for inference, but could be related to using lambda=0 for all obs in the logbook index. Try to increase the lambda to a very small value -- might fix it.